Move commands implementation to the inspektor.commands namespace
This way we have more clarity of where the commands
are implemented.
