+++ inspektor/reindent.py
@@ -16,14 +16,13 @@ import logging
 import os
 import sys
 import tokenize
+
 import six
 
 try:
     from os.path import walk
 except ImportError:
     from os import walk
-
-from cliff.command import Command
 
 from .inspector import PathInspector
 from .utils import stacktrace
@@ -260,42 +259,3 @@ class Reindenter(object):
         else:
             self.log.warning("Invalid location '%s'", path)
             return False
-
-
-class ReindentCommand(Command):
-    """
-    check for correct file indentation
-    """
-    log = logging.getLogger(__name__)
-
-    def get_parser(self, prog_name):
-        parser = super(ReindentCommand, self).get_parser(prog_name)
-        parser.add_argument('path', type=str,
-                            help='Path to check (empty for full tree check)',
-                            nargs='*',
-                            default=None)
-        parser.add_argument('--fix', action='store_true', default=False,
-                            help='Fix any indentation problems found')
-        parser.add_argument('--exclude', type=str,
-                            help='Quoted string containing paths or '
-                                 'patterns to be excluded from '
-                                 'checking, comma separated')
-        parser.add_argument('--verbose', action='store_true',
-                            help='Print extra debug messages')
-        return parser
-
-    def take_action(self, parsed_args):
-        if not parsed_args.path:
-            parsed_args.path = [os.getcwd()]
-
-        reindenter = Reindenter(parsed_args, logger=self.log)
-
-        status = True
-        for path in parsed_args.path:
-            status &= reindenter.check(path)
-        if status:
-            self.log.info('Indentation check PASS')
-            return 0
-        else:
-            self.log.error('Syntax check FAIL')
-            return 1
