+++ inspektor/style.py
@@ -21,7 +21,6 @@ try:
 except ImportError:
     from os import walk
 
-from cliff.command import Command
 import pycodestyle
 
 try:
@@ -111,51 +110,3 @@ class StyleChecker(object):
         else:
             self.log.error("Invalid location '%s'", path)
             return False
-
-
-class StyleCommand(Command):
-    """
-    check against style guide with pycodestyle
-    """
-    log = logging.getLogger(__name__)
-
-    def get_parser(self, prog_name):
-        parser = super(StyleCommand, self).get_parser(prog_name)
-        parser.add_argument('path', type=str,
-                            help='Path to check (empty for full tree check)',
-                            nargs='*',
-                            default=None)
-        parser.add_argument('--disable', type=str,
-                            help='Disable the PEP8 errors given as arguments. '
-                                 'Default: %(default)s',
-                            default='E501,E265,W601,E402')
-        parser.add_argument('--fix', action='store_true', default=False,
-                            help='Fix any style problems found '
-                                 '(needs autopep8 installed)')
-        parser.add_argument('--max-line-length', type=int, default=79,
-                            help=('set maximum allowed line length. Default: '
-                                  '%(default)s'))
-        parser.add_argument('--exclude', type=str,
-                            help='Quoted string containing paths or '
-                                 'patterns to be excluded from '
-                                 'checking, comma separated')
-        parser.add_argument('--verbose', action='store_true',
-                            help='Print extra debug messages')
-        return parser
-
-    def take_action(self, parsed_args):
-        paths = parsed_args.path
-        if not paths:
-            paths = [os.getcwd()]
-
-        style_checker = StyleChecker(parsed_args, logger=self.log)
-
-        status = True
-        for path in paths:
-            status &= style_checker.check(path)
-        if status:
-            self.log.info("PEP8 compliance check PASS")
-            return 0
-        else:
-            self.log.error("PEP8 compliance FAIL")
-            return 1
