+++ inspektor/utils/__init__.py
@@ -15,6 +15,7 @@
 import logging
 import random
 import string
+
 import six
 
 from . import download
