+++ inspektor/utils/download.py
@@ -16,6 +16,7 @@ import logging
 import os
 import shutil
 import socket
+
 import six
 
 # pylint: disable=E0611
